+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### Fixes & tech 🚒
 
+* Include checked/unchecked changes in granular-change-update mechanism (#945)
 * New state handling for errors (observable subject) (#943)
 * Remove flickering effect when opening navigation screen (#941)
 * Disable Crashlytics crash reporting for debug builds (#940)